+++ android/app/src/main/java/com/bghitech/momenta/data/repository/RepositoryModule.kt
@@ -31,7 +31,4 @@ abstract class RepositoryModule {
     @Singleton
     abstract fun bindProfileRepository(impl: ProfileRepositoryImpl): ProfileRepository
 
-    @Binds
-    @Singleton
-    abstract fun bindSettingsRepository(impl: SettingsRepositoryImpl): SettingsRepository
 }